Fix THE END location
The THE END location did not function properly before. Because THE END is not in Panels (it's in Decorations), node lookup did not work right before and the THE END location ended up using the panel from the previous location, which was ENTER (fifth floor orange). This fixes that lookup so that the location works correctly.